dyno tester firmware update for new tester hardware.
Compiles, but untested.

Added Calibration struct for DAC/ADC.
Added calibration to eeprom as separate struct from config, with its own crc.

Added multiple gain ranges for input voltage sampling.
Updated test_sample() to use input gain ranges.
Updated get_sample_voltage() and _tol() for above, and simplifed call.
Updated gain resistor values and pin connections.
Sample read is now its own function, and is used for output return to input check during sweep test.
Gain adjustment routine is now also a separate function, so it can be applied to input or output.

Fixed some minor bugs - command reads until '\n' instead of '/n'
- EEP_PUT/_GET semicolons removed

+++ simple_oscillator/pwm.ino
@@ -446,7 +446,7 @@ void set_pwm_prescale(byte prescale)
     reg &= ~0b111;
     reg |= prescale;
     TCCR1B = reg;
-    byte scale[] = {0,1,8,64,256,1024};
+    unsigned int scale[] = {0,1,8,64,256,1024};
     unsigned int clk[] = {0, 16000, 2000, 250, 63, 16};
     unsigned int frq[] = {0, 62500, 7813, 977, 244, 61};
     /* move this to commands, so that only a command will illicit a response
